pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transactoid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "promptorium-python @ git+https://github.com/adambossy/promptorium-python@main",
    "openai>=1.0.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
    "openai-agents>=0.5.1",
    "pydantic>=2.8",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9.11",
    "alembic>=1.17.2",
    "openai-chatkit>=1.4.1",
    "fastapi>=0.127.0",
    "uvicorn>=0.38.0",
    "loguru>=0.7.3",
    "rich>=13.0.0",
    "resend>=2.0.0",
    "boto3>=1.26.0",
    "google-adk>=1.14.0",
]

[project.optional-dependencies]
stagehand = [
    "stagehand>=0.5.9",
]

[project.scripts]
transactoid = "transactoid.ui.cli:main"

[tool.setuptools]
package-dir = {"transactoid" = "src/transactoid"}

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["transactoid*", "models", "evals*", "scripts"]

[tool.deadcode]
exclude = [
    ".venv/",
    ".venv/**",
    ".worktrees/",
    ".worktrees/**",
    "db/migrations/",
    "db/migrations/**",
    "**/migrations/",
    "**/migrations/**",
    "src/transactoid/adapters/db/models.py",
    "src/transactoid/ui/mcp/server.py",
    "src/transactoid/ui/cli.py",
    "src/transactoid/ui/chatkit/server.py",
    "src/transactoid/ui/chatkit/adapter.py",
    "scripts/run.py",
    "src/transactoid/adapters/storage/",
    "src/transactoid/adapters/storage/**",
    "src/transactoid/tools/storage/",
    "src/transactoid/tools/storage/**",
    "src/transactoid/jobs/report/",
    "src/transactoid/jobs/report/**",
    "evals/core/",
    "evals/core/**",
    "build/",
    "dist/",
    "**/site-packages/",
]
ignore-names = [
    "__all__",
    "__doc__",
    "__path__",
    "__version__",
    "_",
    "logger",
    "LOGGER",
    "*_fixture",
    "created_at",
    "updated_at",
    "input_schema",
    "confidence_level",
    "detailed",
    "primary",
    "enum",
    "properties",
    "required",
]
ignore-names-in-files = [
    "*/__init__.py",
    "tests/**",
    "test/**",
]
ignore-definitions = [
    "Migration",
    "*Migration",
    "RecategorizeTool",
    "TagTransactionsTool",
    "create_context",
]
ignore-definitions-if-inherits-from = [
    "BaseModel",
    "Schema",
    "Settings",
    "Protocol",
    "BaseHTTPRequestHandler",
    "_StoreBase",
]
ignore-definitions-if-decorated-with = [
    "mcp.tool",
    "mcp.prompt",
    "app.command",
    "run_app.command",
    "app.callback",
    "app.get",
    "app.post",
    "function_tool",
]

[dependency-groups]
dev = [
    "deadcode>=2.4.1",
    "mypy>=1.18.2",
    "pytest>=9.0.0",
    "ruff>=0.14.4",
]
